wip: Simplifying coder to a single class
